Deleted old input handling code
